+++ openerp_proxy/ext/repr.py
@@ -16,6 +16,7 @@ from openerp_proxy.orm.record import Record
 from openerp_proxy.orm.object import Object
 from openerp_proxy.core import Client
 from openerp_proxy.utils import AttrDict
+from openerp_proxy.session import Session
 
 from IPython.display import HTML, FileLink
 
@@ -621,3 +622,37 @@ class ClientHTML(Client):
 
         table = ttable % (caption + data)
         return html % (table + help_text)
+
+
+class IPYSession(Session):
+    def _repr_html_(self):
+        """ Provides HTML representation of session (Used for IPython)
+        """
+        from openerp_proxy.utils import ustr as _
+
+        def _get_data():
+            for url in self._databases.keys():
+                index = self._index_url(url)
+                aliases = (_(al) for al, aurl in self.aliases.items() if aurl == url)
+                yield (url, index, u", ".join(aliases))
+        ttable = u"<table style='display:inline-block'>%s</table>"
+        trow = u"<tr>%s</tr>"
+        tdata = u"<td>%s</td>"
+        caption = u"<caption>Previous connections</caption>"
+        hrow = u"<tr><th>DB URL</th><th>DB Index</th><th>DB Aliases</th></tr>"
+        help_text = (u"<div style='display:inline-block;vertical-align:top;margin-left:10px;'>"
+                     u"To get connection just call<br/> <ul>"
+                     u"<li>session.<b>aliase</b></li>"
+                     u"<li>session[<b>index</b>]</li>"
+                     u"<li>session[<b>aliase</b>]</li> "
+                     u"<li>session[<b>url</b>]</li>"
+                     u"<li>session.get_db(<b>url</b>|<b>index</b>|<b>aliase</b>)</li>"
+                     u"</ul></div>")
+
+        data = u""
+        for row in _get_data():
+            data += trow % (u''.join((tdata % i for i in row)))
+
+        table = ttable % (caption + hrow + data)
+
+        return u"<div>%s %s</div>" % (table, help_text)